- run python-modernize on repo - replace smart_unicode w smart_text - fix print( , file=) in comment - git tag v1.0.0-yr.modernize

+++ piston/models.py
@@ -1,3 +1,4 @@
+from __future__ import absolute_import
 import urllib, time, urlparse
 
 # Django imports
@@ -7,8 +8,8 @@ from django.contrib.auth.models import User
 from django.core.mail import send_mail, mail_admins
 
 # Piston imports
-from managers import TokenManager, ConsumerManager, ResourceManager
-from signals import consumer_post_save, consumer_post_delete
+from .managers import TokenManager, ConsumerManager, ResourceManager
+from .signals import consumer_post_save, consumer_post_delete
 
 KEY_SIZE = 18
 SECRET_SIZE = 32
@@ -85,7 +86,7 @@ class Token(models.Model):
     secret = models.CharField(max_length=SECRET_SIZE)
     verifier = models.CharField(max_length=VERIFIER_SIZE)
     token_type = models.IntegerField(choices=TOKEN_TYPES)
-    timestamp = models.IntegerField(default=long(time.time()))
+    timestamp = models.IntegerField(default=int(time.time()))
     is_approved = models.BooleanField(default=False)
 
     user = models.ForeignKey(User, null=True, blank=True, related_name='tokens')
